# OPENAPI_DRAFT_V0.yaml
openapi: 3.1.0
info:
  title: MahjongLab API
  version: 0.1.0
  summary: Review-first API draft for MahjongLab
  description: |
    This draft is aligned with the current prototype migration checklist.
    Priority is:
    1. Review flow
    2. Match flow
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - BearerAuth: []
paths:
  /api/me:
    get:
      tags: [auth]
      summary: Get current user profile
      operationId: getMe
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/dashboard/summary:
    get:
      tags: [dashboard]
      summary: Get dashboard summary for home page
      operationId: getDashboardSummary
      responses:
        '200':
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/uploads:
    post:
      tags: [uploads]
      summary: Upload a replay file
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Upload success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/review-jobs:
    post:
      tags: [review-jobs]
      summary: Create a review job
      operationId: createReviewJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewJobRequest'
      responses:
        '201':
          description: Review job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/review-jobs/{task_id}:
    get:
      tags: [review-jobs]
      summary: Get a review job status
      operationId: getReviewJob
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Review job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/review-jobs/{task_id}/result:
    get:
      tags: [review-jobs]
      summary: Get result mapping for a review job
      operationId: getReviewJobResult
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Review job result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewJobResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/review-jobs/{task_id}/retry:
    post:
      tags: [review-jobs]
      summary: Retry a failed review job
      operationId: retryReviewJob
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Review job re-queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/reviews:
    get:
      tags: [reviews]
      summary: List reviews
      operationId: listReviews
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Q'
        - name: platform
          in: query
          schema:
            type: string
            enum: [all, tenhou, majsoul, internal]
        - name: date_range
          in: query
          schema:
            type: string
            enum: [all, today, week, month]
      responses:
        '200':
          description: Review list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/reviews/{review_id}:
    get:
      tags: [reviews]
      summary: Get one review summary
      operationId: getReview
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '200':
          description: Review summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [reviews]
      summary: Delete a review
      operationId: deleteReview
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '204':
          description: Review deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/reviews/{review_id}/entries:
    get:
      tags: [reviews]
      summary: List review entries
      operationId: listReviewEntries
      parameters:
        - $ref: '#/components/parameters/ReviewId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: kyoku
          in: query
          schema:
            type: integer
            minimum: 0
        - name: deviation_level
          in: query
          schema:
            type: string
            enum: [all, none, low, medium, high]
        - name: decision_type
          in: query
          schema:
            type: string
            enum: [all, discard, riichi, chi, pon, kan, agari, ryukyoku, pass, other]
      responses:
        '200':
          description: Review entry list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewEntryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReviewId:
      name: review_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Q:
      name: q
      in: query
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    UserProfile:
      type: object
      required: [id, display_name, locale, timezone, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
        locale:
          type: string
        timezone:
          type: string
        created_at:
          type: string
          format: date-time
    DashboardSummary:
      type: object
      required:
        - review_count
        - match_count
        - mistake_count
        - recent_reviews
        - recent_matches
      properties:
        review_count:
          type: integer
        match_count:
          type: integer
        mistake_count:
          type: integer
        recent_reviews:
          type: array
          items:
            $ref: '#/components/schemas/ReviewListItem'
        recent_matches:
          type: array
          items:
            $ref: '#/components/schemas/MatchListItem'
...
    UploadResponse:
      type: object
      required: [file_key, original_filename, content_type, size]
      properties:
        file_key:
          type: string
        original_filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
    CreateReviewJobRequest:
      type: object
      required: [source_type, source]
      properties:
        source_type:
          type: string
          enum: [internal_match, tenhou_url, tenhou_id, majsoul_url, upload_file, inline_json]
        platform:
          type: string
          nullable: true
          enum: [tenhou, majsoul, internal, null]
        source:
          type: object
          additionalProperties: true
          description: |
            Source payload. Example:
            - {"url":"https://..."}
            - {"external_id":"..."}
            - {"file_key":"uploads/..."}
            - {"events":[...]}
        target_player_ref:
          type: string
          nullable: true
        kyoku_filter:
          type: string
          default: all
        lang:
          type: string
          default: zh-CN
        anonymous:
          type: boolean
          default: false
        output_format:
          type: string
          enum: [json, html, both]
          default: json
    ReviewJob:
      type: object
      required:
        - id
        - status
        - progress
        - step
        - source_type
        - source
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [created, parsing, queued, analyzing, completed, failed, cancelled]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        step:
          type: string
        source_type:
          type: string
          enum: [internal_match, tenhou_url, tenhou_id, majsoul_url, upload_file, inline_json]
        platform:
          type: string
          nullable: true
        source:
          type: object
          additionalProperties: true
        target_player_ref:
          type: string
          nullable: true
        target_actor:
          type: integer
          minimum: 0
          maximum: 3
          nullable: true
        review_id:
          type: string
          format: uuid
          nullable: true
        error_code:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        attempt_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        queued_at:
          type: string
          format: date-time
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    ReviewJobResult:
      type: object
      required: [task_id, status]
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, parsing, queued, analyzing, completed, failed, cancelled]
        review_id:
          type: string
          format: uuid
          nullable: true
    Review:
      type: object
      required:
        - id
        - job_id
        - user_id
        - target_actor
        - engine_name
        - engine_version
        - reviewed_decision_count
        - match_decision_count
        - high_deviation_count
        - medium_deviation_count
        - optimal_count
        - summary
        - stats
        - created_at
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
          nullable: true
        platform:
          type: string
          nullable: true
        target_actor:
          type: integer
          minimum: 0
          maximum: 3
        target_player_label:
          type: string
          nullable: true
        engine_name:
          type: string
        engine_version:
          type: string
        model_tag:
          type: string
          nullable: true
        lang:
          type: string
        reviewed_decision_count:
          type: integer
        match_decision_count:
          type: integer
        high_deviation_count:
          type: integer
        medium_deviation_count:
          type: integer
        optimal_count:
          type: integer
        rating:
          type: number
          nullable: true
        temperature:
          type: number
          nullable: true
        summary:
          type: object
          additionalProperties: true
        stats:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    ReviewListItem:
      allOf:
        - $ref: '#/components/schemas/Review'
    ReviewListResponse:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewListItem'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    ReviewEntry:
      type: object
      required:
        - id
        - review_id
        - seq
        - kyoku_index
        - honba
        - junme
        - tiles_left
        - decision_type
        - actual_action
        - expected_action
        - is_match
        - deviation_level
        - details
      properties:
        id:
          type: integer
          format: int64
        review_id:
          type: string
          format: uuid
        seq:
          type: integer
        kyoku_index:
          type: integer
        honba:
          type: integer
        junme:
          type: integer
        tiles_left:
          type: integer
        last_actor:
          type: integer
          minimum: 0
          maximum: 3
          nullable: true
        tile:
          type: string
          nullable: true
        decision_type:
          type: string
          enum: [discard, riichi, chi, pon, kan, agari, ryukyoku, pass, other]
        actual_action:
          type: object
          additionalProperties: true
        expected_action:
          type: object
          additionalProperties: true
        is_match:
          type: boolean
        deviation_level:
          type: string
          enum: [none, low, medium, high]
        delta_score:
          type: integer
          nullable: true
        shanten:
          type: integer
          nullable: true
        at_furiten:
          type: boolean
          nullable: true
        details:
          type: array
          items:
            type: object
            additionalProperties: true
        state_snapshot:
          type: object
          additionalProperties: true
          nullable: true
        tags:
          type: array
          items:
            type: string
    ReviewEntryListResponse:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewEntry'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
    MatchListItem:
      type: object
      required:
        - id
        - created_at
        - match_type
        - difficulty
        - user_rank
        - user_delta
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        match_type:
          type: string
          enum: [tonpu, hanchan]
        difficulty:
          type: string
        training_goal:
          type: string
          nullable: true
        user_rank:
          type: integer
        user_delta:
          type: integer
        review_job_id:
          type: string
          format: uuid
          nullable: true
...